requirements.txt
pywin32>=306
pyautogui>=0.9.54
mss>=9.0.1
Pillow>=10.3.0
PyYAML>=6.0.2
pytest>=8.3.3
pytest-html>=4.1.1
colorama>=0.4.6
pyperclip>=1.8.2
pytesseract>=0.3.10
opencv-python>=4.10.0.84
numpy>=1.26.0